+++ src/main/java/com/pu/carmanagment/Repository/MaintenanceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaintenanceRepository extends JpaRepository<Maintenance,Long> {
+public interface MaintenanceRepository extends JpaRepository<Maintenance,Integer> {
 }
